Beta version
Looks at the inbox, processes the e-mails and moves them into another
folder when done. The resulting information is appended to a local DB
(SQLite). Whenever desired the information in the DB can be accessed via
a webpage, showing a table of all clients and their status in the last N
days:

http://my-server:5009/status?days=100

Every day, the emails are processed at 11:00, local time, and a ticket
is sent when done.